[cmake] Don't impose CMAKE_BUILD_TYPE and debug flags on other projects
The framework exports some cmake project files to make it easier to
import its targets and settings from other projects depending on it
(e.g. plugins). This is fine and very useful.

However, the exported project files are forcing the CMAKE_BUILD_TYPE
that was configured when building the framework to other projects, thus
preventing the other projects to use a CMAKE_BUILD_TYPE on their builds
that is different and independent from the one used when building the
framework. The same is happening for debug related flags like NDEBUG,
etc. This is not desired, so fix it.

Remove all settings for CMAKE_BUILD_TYPE from the exported cmake files
and instead do them only for the framework the its source's top-level
cmake file.

To avoid debug related and other flags contamination on other projects,
also move these settings from the 'CompileSettings' target into the new
'CompileSettingsDebug' target and add it as a private dependency to the
framework targets, so those settings still get applied here but do not
automatically contaminate other external targets that depend on this
one. If those targets want the same flags, they can always add a
dependency to 'CompileSettingsDebug' as well, as this new target is
also exported now.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>